+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## [Unreleased]
 ### Changed
+- `StepReporter.sendStep` methods now returns `Maybe<String>` with Step ID, by @HardNorth
+
+## [5.2.25]
+### Changed
 - `StepReporter.step(java.lang.String)` method now returns `Maybe<String>` with Step ID, by @HardNorth
+- `jackson-databind` version updated on 2.15.4, by @HardNorth
 
 ## [5.2.24]
 ### Fixed